backend/requirements.txt
aiohttp==3.9.5
aioice==0.9.0
aiortc==1.9.0
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
attrs==23.2.0
av==12.2.0
blinker==1.8.2
cachetools==5.3.3
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.8
decorator==4.4.2
distro==1.9.0
dnspython==2.6.1
email_validator==2.2.0
exceptiongroup==1.2.1
fastapi==0.111.0
fastapi-cli==0.0.4
frozenlist==1.4.1
gitdb==4.0.11
GitPython==3.1.43
google-crc32c==1.5.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
ifaddr==0.2.0
imageio==2.34.2
imageio-ffmpeg==0.5.1
importlib_resources==6.4.0
install==1.3.5
Jinja2==3.1.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
moviepy==1.0.3
multidict==6.0.5
numpy==1.24.4
openai==1.31.0
opencv-python==4.10.0.84
orjson==3.10.6
packaging==24.1
pandas==2.0.3
pillow==10.3.0
pkgutil_resolve_name==1.3.10
proglog==0.1.10
protobuf==5.27.2
pyarrow==16.1.0
PyAudio==0.2.14
pycparser==2.22
pydantic==2.7.3
pydantic_core==2.18.4
pydeck==0.9.1
pydub==0.25.1
pyee==11.1.0
pygame==2.5.2
Pygments==2.18.0
pylibsrtp==0.10.0
pyOpenSSL==24.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
requests==2.32.3
rich==13.7.1
rpds-py==0.18.1
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
SpeechRecognition==3.10.4
starlette==0.37.2
streamlit==1.36.0
streamlit-audiorec==0.1.3
streamlit-webrtc==0.47.7
tenacity==8.4.2
toml==0.10.2
toolz==0.12.1
tornado==6.4.1
tqdm==4.66.4
typer==0.12.3
typing_extensions==4.12.1
tzdata==2024.1
ujson==5.10.0
urllib3==2.2.1
uvicorn==0.30.1
uvloop==0.19.0
watchdog==4.0.1
watchfiles==0.22.0
websockets==12.0
yarl==1.9.4
zipp==3.19.2
